Fix compilation on Windows: - Fixes for type-conversion   (time_t   is not interchangeable  with  my_time_t on Windows as time_t s 64 bit while my_time_t is long) - BIGENDIAN-> ARCH_BIGENDIAN .   BIGENDIAN constant  is defined in winsock2.h (as 0) - added explicit cast for longlong->double conversion in sql/item.h  (fixed many warnings)
Also, HAVE_SNPRINTF is now defined and snprintf is defined to _snprintf in config-win.h

+++ sql-common/my_time.c
@@ -1245,7 +1245,7 @@ int number_to_time(double nr, MYSQL_TIME *ltime, int *was_cut)
   ltime->hour  = tmp/100/100;
   ltime->minute= tmp/100%100;
   ltime->second= tmp%100;
-  ltime->second_part= (nr-tmp)*1e6;
+  ltime->second_part= (ulong)((nr-tmp)*1e6);
 
   if (ltime->minute < 60 && ltime->second < 60)
     return 0;
@@ -1363,7 +1363,7 @@ MYSQL_TIME *unpack_time(longlong packed, MYSQL_TIME *my_time)
   get_one(my_time->hour,             24ULL);
   get_one(my_time->day,              32ULL);
   get_one(my_time->month,            13ULL);
-  my_time->year= packed;
+  my_time->year= (uint)packed;
   my_time->time_type= MYSQL_TIMESTAMP_DATETIME;
   return my_time;
 }
